// packages/ESRouter/src/matcher/types.ts
import {
  RouteRecordMultipleViews,
  NavigationGuard,
  _RouteRecordBase,
  _RouteRecordProps,
  NavigationGuardNextCallback,
  RouteRecordRaw,
} from '../types'
import {ComponentPublicInstance} from 'vue'

// normalize component/components into components and make every property always present
/**
 * Normalized version of a {@link RouteRecord | route record}.
 */
export interface RouteRecordNormalized {
  /**
   * {@inheritDoc _RouteRecordBase.path}
   */
  path: _RouteRecordBase['path']
  /**
   * {@inheritDoc _RouteRecordBase.redirect}
   */
  redirect: _RouteRecordBase['redirect'] | undefined
  /**
   * {@inheritDoc _RouteRecordBase.name}
   */
  name: _RouteRecordBase['name']
  /**
   * {@inheritDoc RouteRecordMultipleViews.components}
   */
  components: RouteRecordMultipleViews['components'] | null | undefined
  /**
   * Nested route records.
   */
  children: RouteRecordRaw[]
  /**
   * {@inheritDoc _RouteRecordBase.meta}
   */
  meta: Exclude<_RouteRecordBase['meta'], void>
  /**
   * {@inheritDoc RouteRecordMultipleViews.props}
   */
  props: Record<string, _RouteRecordProps>
  /**
   * Registered beforeEnter guards
   */
  beforeEnter: _RouteRecordBase['beforeEnter']
  /**
   * Registered leave guards
   *
   * @internal
   */
  leaveGuards: Set<NavigationGuard>
  /**
   * Registered update guards
   *
   * @internal
   */
  updateGuards: Set<NavigationGuard>
  /**
   * Registered beforeRouteEnter callbacks passed to `next` or returned in guards
   *
   * @internal
   */
  enterCallbacks: Record<string, NavigationGuardNextCallback[]>
  /**
   * Mounted route component instances
   * Having the instances on the record mean beforeRouteUpdate and
   * beforeRouteLeave guards can only be invoked with the latest mounted app
   * instance if there are multiple application instances rendering the same
   * view, basically duplicating the content on the page, which shouldn't happen
   * in practice. It will work if multiple apps are rendering different named
   * views.
   */
  instances: Record<string, ComponentPublicInstance | undefined | null>
  // can only be of the same type as this record
  /**
   * Defines if this record is the alias of another one. This property is
   * `undefined` if the record is the original one.
   */
  aliasOf: RouteRecordNormalized | undefined

  /**
   *  standard | singleTask | clearTask
   */
  launchMode?: ESRouteLaunchMode

  /**
   * lifecycle
   */
  lifecycle: LifecycleState

  /**
   * type
   */
  type?: ESRouteType
}

/**
 * {@inheritDoc RouteRecordNormalized}
 */
export type RouteRecord = RouteRecordNormalized

/**
 *
 */
export enum ESRouteLaunchMode {
  ES_ROUTE_LAUNCH_MODE_CLEAR_TASK = 'clearTask',
  ES_ROUTE_LAUNCH_MODE_SINGLE_TASK = 'singleTask',
  ES_ROUTE_LAUNCH_MODE_STANDARD = 'standard',
}

export enum ESRouteType {
  ES_ROUTE_TYPE_PAGE = 0,
  ES_ROUTE_TYPE_DIALOG = 1,
  ES_ROUTE_TYPE_SLOT = 2,
}

export enum LifecycleState {
  uninitialized = -1,
  initialized = 0,
  create = 1,
  restart = 2,
  start = 3,
  restoreInstanceState = 4,
  resume = 5,
  pause = 6,
  stop = 7,
  saveInstanceState = 8,
  destroy = 9,
}